feat: implement update, create and delete on repository

+++ repositories/post/drizzle-post.repository.ts
@@ -44,9 +44,69 @@ export class DrizzleRepository implements IPostRepository {
     return post;
   }
 
-  async deleteById(id: string): Promise<void> {
+  async create(post: PostModel): Promise<void> {
+    const postExists = await drizzleDb.query.posts.findFirst({
+      where: (posts, { or, eq }) =>
+        or(eq(posts.id, post.id), eq(posts.slug, post.slug)),
+      columns: { id: true },
+    });
+    if (!!postExists) {
+      throw new Error('Já existe um post com esse id ou slug');
+    }
+    try {
+      await drizzleDb.insert(postsTable).values(post);
+    } catch (e) {
+      throw new Error('[ERR-001]: Por favor, contate o suporte');
+    }
+  }
+
+  async update(
+    id: string,
+    newPost: Omit<PostModel, 'id' | 'slug' | 'createdAt' | 'updatedAt'>,
+  ): Promise<PostModel> {
+    const oldPost = await drizzleDb.query.posts.findFirst({
+      where: (post, { eq }) => eq(post.id, id),
+    });
+    if (!oldPost) {
+      throw new Error('Post não encontrado na base de dados');
+    }
+
+    const now = new Date().toISOString();
+    const postData = {
+      title: newPost.title,
+      excerpt: newPost.excerpt,
+      content: newPost.content,
+      coverImageUrl: newPost.coverImageUrl,
+      published: newPost.published,
+      updatedAt: now,
+      author: newPost.author,
+    };
+
+    try {
+      await drizzleDb
+        .update(postsTable)
+        .set(postData)
+        .where(eq(postsTable.id, id));
+      return { ...oldPost, ...postData };
+    } catch (e) {
+      throw new Error('[ERR-006]: Erro desconhecido');
+    }
+  }
+
+  async deleteById(id: string): Promise<PostModel> {
     formatLog('deleteById');
-    await drizzleDb.delete(postsTable).where(eq(postsTable.id, id));
+    const post = await drizzleDb.query.posts.findFirst({
+      where: (posts, { eq }) => eq(posts.id, id),
+    });
+    if (!post) {
+      throw new Error('Post não existe na base de dados');
+    }
+    try {
+      await drizzleDb.delete(postsTable).where(eq(postsTable.id, id));
+      return post;
+    } catch (e) {
+      throw new Error('[ERR-002]: Por favor, contate o suporte');
+    }
   }
 }
